+++ .storybook/preview.tsx
@@ -1,0 +1,25 @@
+import React from 'react'
+import { Preview } from '@storybook/react'
+import { ThemeProvider } from '@mui/system'
+import theme from '../src/lib/theme'
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+  decorators: [
+    (Story) => (
+      <ThemeProvider theme={theme}>
+        <Story />
+      </ThemeProvider>
+    ),
+  ],
+}
+
+export default preview
