+++ src/lib/theme.ts
@@ -1,3 +1,4 @@
+// theme.ts
 import { createTheme } from '@mui/material/styles'
 
 const theme = createTheme({
